feat: post readings to api for long-term storage

+++ backend/models.py
@@ -9,7 +9,7 @@ class Token(BaseModel):
 
 class Reading(BaseModel):
     timestamp: datetime.datetime
-    value: int
+    value: float
 
 
 class ReadingResponse(BaseModel):
